+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/internal/ModulesSupport.java
@@ -1,14 +1,24 @@
 package ru.vyarus.dropwizard.guice.module.installer.internal;
 
+import com.google.inject.Binding;
+import com.google.inject.Key;
 import com.google.inject.Module;
+import com.google.inject.spi.Element;
+import com.google.inject.spi.Elements;
 import com.google.inject.util.Modules;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ru.vyarus.dropwizard.guice.GuiceyOptions;
+import ru.vyarus.dropwizard.guice.module.GuiceBootstrapModule;
 import ru.vyarus.dropwizard.guice.module.context.ConfigurationContext;
 import ru.vyarus.dropwizard.guice.module.context.option.Options;
+import ru.vyarus.dropwizard.guice.module.installer.util.BindingUtils;
 import ru.vyarus.dropwizard.guice.module.support.*;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
+
+import static ru.vyarus.dropwizard.guice.GuiceyOptions.ConfigureFromGuiceModules;
+import static ru.vyarus.dropwizard.guice.GuiceyOptions.InjectorStage;
 
 /**
  * Helper class for guice modules processing.
@@ -17,6 +27,8 @@ import java.util.List;
  * @since 25.04.2018
  */
 public final class ModulesSupport {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(ModulesSupport.class);
 
     private ModulesSupport() {
     }
@@ -55,14 +67,80 @@ public final class ModulesSupport {
      * @return modules for injector creation
      */
     public static Iterable<Module> prepareModules(final ConfigurationContext context) {
-        final List<Module> normalModules = context.getNormalModules();
+        List<Module> normalModules = context.getNormalModules();
         final List<Module> overridingModules = context.getOverridingModules();
         // use different lists to avoid possible side effects from listeners (not allowed to exclude or modify order)
         context.lifecycle().injectorCreation(
                 new ArrayList<>(normalModules),
                 new ArrayList<>(overridingModules),
                 context.getDisabledModules());
+
+        // repackage normal modules to reveal all guice extensions
+        normalModules = analyzeModules(context);
         return overridingModules.isEmpty() ? normalModules
                 : Collections.singletonList(Modules.override(normalModules).with(overridingModules));
+    }
+
+    /**
+     * Search for extensions in guice bindings (directly declared in modules).
+     * Only user provided modules are analyzed. Overriding modules are not analyzed.
+     * <p>
+     * Use guice SPI. In order to avoid duplicate analysis in injector creation time, wrap
+     * parsed elements as new module (and use it instead of original modules). Also, if
+     * bound extension is disabled, target binding is simply removed (in order to
+     * provide the same disable semantic as with usual extensions).
+     *
+     * @param context configuration context
+     * @return list of repackaged modules to use
+     */
+    private static List<Module> analyzeModules(final ConfigurationContext context) {
+        List<Module> modules = context.getNormalModules();
+        final Boolean configureFromGuice = context.option(ConfigureFromGuiceModules);
+        // one module mean no user modules registered
+        if (modules.size() > 1 && configureFromGuice) {
+            // analyzing only user bindings (excluding overrides and guicey technical bindings)
+            final GuiceBootstrapModule bootstrap = (GuiceBootstrapModule) modules.remove(modules.size() - 1);
+            try {
+                // find extensions and remove bindings if required (disabled extensions)
+                final List<Element> elements = analyzeBindings(context, modules);
+                // wrap raw elements into module to avoid duplicate work on guice startup and put back bootstrap
+                modules = Arrays.asList(Elements.getModule(elements), bootstrap);
+            } catch (Exception ex) {
+                // better show meaningful message then just fail entire startup with ambiguous message
+                // NOTE if guice configuration is not OK it will fail here too, but user will see injector creation
+                // error as last error in logs.
+                LOGGER.error("Failed to analyze guice bindings - skipping this step. Note that configuration"
+                        + " from bindings may be switched off with " + GuiceyOptions.class.getSimpleName() + "."
+                        + ConfigureFromGuiceModules.name() + " option.", ex);
+                // recover and use original modules
+                modules.add(bootstrap);
+            }
+        }
+        return modules;
+    }
+
+    private static List<Element> analyzeBindings(final ConfigurationContext context, final List<Module> modules) {
+        final List<Element> elements = Elements.getElements(context.option(InjectorStage), modules);
+        final Iterator<Element> it = elements.iterator();
+        while (it.hasNext()) {
+            final Element element = it.next();
+            // filter constants, listeners, aop etc.
+            if (element instanceof Binding) {
+                final Key key = ((Binding) element).getKey();
+                // extensions bindings may be only unqualified, class only (no generified types)
+                if (key.getAnnotation() == null && key.getTypeLiteral().getType() instanceof Class) {
+                    final Class type = key.getTypeLiteral().getRawType();
+                    if (ExtensionsSupport.registerExtensionBinding(context, type,
+                            BindingUtils.getTopDeclarationModule(element))) {
+                        LOGGER.debug("Extension detected from guice binding: {}", type.getSimpleName());
+                        if (!context.isExtensionEnabled(type)) {
+                            it.remove();
+                            LOGGER.debug("Removed disabled extension binding: {}", type.getSimpleName());
+                        }
+                    }
+                }
+            }
+        }
+        return elements;
     }
 }