+++ CHANGELOG.md
@@ -82,6 +82,15 @@
 * (breaking) `DebugGuiceyLifecycle` listener renamed into `LifecycleDiagnostic`
 * (breaking) All diagnostic reports moved into one top-level "debug" package.
 * Add guice aop appliance report
+* Support extensions recognition from guice modules: 
+    - extensions are detected from declaration in specified guice modules 
+        (essentially same as classpath scan, but from bindings)
+    - support only direct type bindings (all generified or qualified declarations ignored)    
+    - all extension registration types may work together (classpath scan, manual declaration and binding declaration)    
+    - extensions registered directly (or found by classpath scan) and also bound manually in guice module 
+        will not conflict anymore (as manual declaration would be detected) and so @LazyBinding workaround is not needed
+    - enabled by default, but can be disabled with `GuiceyOptions.ConfigureFromGuiceModules`
+    - extensions declared in guice module may be also disabled (guicey will remove binding declaration in this case)    
 * Fix configuration bindings for recursive configuration object declarations (#60)
 
 Main breaking changes were caused by: